merge

+++ src/game/WorldHandlers/MoveMap.h
@@ -2,7 +2,7 @@
  * MaNGOS is a full featured server for World of Warcraft, supporting
  * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
  *
- * Copyright (C) 2005-2017  MaNGOS project <https://getmangos.eu>
+ * Copyright (C) 2005-2021 MaNGOS <https://getmangos.eu>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -29,12 +29,13 @@
 #include "../../dep/recastnavigation/Detour/Include/DetourNavMesh.h"
 #include "../../dep/recastnavigation/Detour/Include/DetourNavMeshQuery.h"
 
+#include "Platform/Define.h"
 #include "Utilities/UnorderedMapSet.h"
 
 class Unit;
 
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
     return (void*)new unsigned char[size];
 }
@@ -57,10 +58,14 @@ namespace MMAP
         ~MMapData()
         {
             for (NavMeshQuerySet::iterator i = navMeshQueries.begin(); i != navMeshQueries.end(); ++i)
-                { dtFreeNavMeshQuery(i->second); }
+            {
+                dtFreeNavMeshQuery(i->second);
+            }
 
             if (navMesh)
-                { dtFreeNavMesh(navMesh); }
+            {
+                dtFreeNavMesh(navMesh);
+            }
         }
 
         dtNavMesh* navMesh;
